requirements.txt
asgiref==3.7.2
Django==5.0
django-bootstrap5==23.4
django-tinymce==3.6.1
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
